services/backend/requirements.txt
# ===== CORE =====
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.3
anyio==3.7.0

# ===== DATABASE =====
asyncpg==0.28.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
redis==4.6.0
alembic==1.12.0

# ===== AI SERVICES =====
openai==1.3.6
elevenlabs==0.2.12
# Current CPU-compatible versions
torch==2.2.1+cpu --index-url https://download.pytorch.org/whl/cpu
torchvision==0.17.1+cpu
torchaudio==2.2.1+cpu

# ===== VECTOR DB =====
pinecone-client==2.2.4

# ===== WEBRTC/MEDIA =====
websockets==11.0.3
python-socketio==5.8.0
aiortc==1.4.0

# ===== UTILITIES =====
httpx==0.25.0
python-dotenv==1.0.0
python-multipart==0.0.6
prometheus-client==0.17.1